server/requirements.txt
Flask==3.0.2
gunicorn==22.0.0
flask-cors==4.0.0
firebase-admin==6.3.0
python-dotenv==1.0.1
langchain-google-genai>=0.0.1
langchain-core>=0.2.0
pydantic>=2.0.0,<3.0.0
protobuf>=3.19.5,<5.0.0
requests==2.31.0
python-jose==3.3.0
PyJWT==2.8.0